+++ tests/examples/restart_deferred_removed/plan1.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+from stepup.core.api import copy, glob
+
+glob("static/**", _defer=True)
+copy("static/foo.txt", "./")
+copy("static/bar.txt", "./")