+++ tests/examples/restart_deferred_removed/plan2.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+from stepup.core.api import copy, glob
+
+glob("static/**", _defer=True)
+copy("static/bar.txt", "./")